# backend/internal/api/rest/ogen/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package ogen

import (
	"fmt"

	"github.com/google/uuid"
)

func (s *ErrorStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// Merged schema.
// Ref: #/components/schemas/Book
type Book struct {
	Author          string    `json:"author"`
	Name            string    `json:"name"`
	Year            int       `json:"year"`
	Price           int64     `json:"price"`
	PublisherId     uuid.UUID `json:"publisherId"`
	PublicationYear int       `json:"publicationYear"`
	Quantity        int64     `json:"quantity"`
	ID              uuid.UUID `json:"id"`
}

// GetAuthor returns the value of Author.
func (s *Book) GetAuthor() string {
	return s.Author
}

// GetName returns the value of Name.
func (s *Book) GetName() string {
	return s.Name
}

// GetYear returns the value of Year.
func (s *Book) GetYear() int {
	return s.Year
}

// GetPrice returns the value of Price.
func (s *Book) GetPrice() int64 {
	return s.Price
}

// GetPublisherId returns the value of PublisherId.
func (s *Book) GetPublisherId() uuid.UUID {
	return s.PublisherId
}

// GetPublicationYear returns the value of PublicationYear.
func (s *Book) GetPublicationYear() int {
	return s.PublicationYear
}

// GetQuantity returns the value of Quantity.
func (s *Book) GetQuantity() int64 {
	return s.Quantity
}

// GetID returns the value of ID.
func (s *Book) GetID() uuid.UUID {
	return s.ID
}

// SetAuthor sets the value of Author.
func (s *Book) SetAuthor(val string) {
	s.Author = val
}

// SetName sets the value of Name.
func (s *Book) SetName(val string) {
	s.Name = val
}

// SetYear sets the value of Year.
func (s *Book) SetYear(val int) {
	s.Year = val
}

// SetPrice sets the value of Price.
func (s *Book) SetPrice(val int64) {
	s.Price = val
}

// SetPublisherId sets the value of PublisherId.
func (s *Book) SetPublisherId(val uuid.UUID) {
	s.PublisherId = val
}

// SetPublicationYear sets the value of PublicationYear.
func (s *Book) SetPublicationYear(val int) {
	s.PublicationYear = val
}

// SetQuantity sets the value of Quantity.
func (s *Book) SetQuantity(val int64) {
	s.Quantity = val
}

// SetID sets the value of ID.
func (s *Book) SetID(val uuid.UUID) {
	s.ID = val
}

// Ref: #/components/schemas/BookInput
type BookInput struct {
	Author          string    `json:"author"`
	Name            string    `json:"name"`
	Year            int       `json:"year"`
	Price           int64     `json:"price"`
	PublisherId     uuid.UUID `json:"publisherId"`
	PublicationYear int       `json:"publicationYear"`
	Quantity        int64     `json:"quantity"`
}

// GetAuthor returns the value of Author.
func (s *BookInput) GetAuthor() string {
	return s.Author
}

// GetName returns the value of Name.
func (s *BookInput) GetName() string {
	return s.Name
}

// GetYear returns the value of Year.
func (s *BookInput) GetYear() int {
	return s.Year
}

// GetPrice returns the value of Price.
func (s *BookInput) GetPrice() int64 {
	return s.Price
}

// GetPublisherId returns the value of PublisherId.
func (s *BookInput) GetPublisherId() uuid.UUID {
	return s.PublisherId
}

// GetPublicationYear returns the value of PublicationYear.
func (s *BookInput) GetPublicationYear() int {
	return s.PublicationYear
}

// GetQuantity returns the value of Quantity.
func (s *BookInput) GetQuantity() int64 {
	return s.Quantity
}

// SetAuthor sets the value of Author.
func (s *BookInput) SetAuthor(val string) {
	s.Author = val
}

// SetName sets the value of Name.
func (s *BookInput) SetName(val string) {
	s.Name = val
}

// SetYear sets the value of Year.
func (s *BookInput) SetYear(val int) {
	s.Year = val
}

// SetPrice sets the value of Price.
func (s *BookInput) SetPrice(val int64) {
	s.Price = val
}

// SetPublisherId sets the value of PublisherId.
func (s *BookInput) SetPublisherId(val uuid.UUID) {
	s.PublisherId = val
}

// SetPublicationYear sets the value of PublicationYear.
func (s *BookInput) SetPublicationYear(val int) {
	s.PublicationYear = val
}

// SetQuantity sets the value of Quantity.
func (s *BookInput) SetQuantity(val int64) {
	s.Quantity = val
}

// Ref: #/components/schemas/Deleted
type Deleted struct {
	Status string `json:"status"`
}

// GetStatus returns the value of Status.
func (s *Deleted) GetStatus() string {
	return s.Status
}

// SetStatus sets the value of Status.
func (s *Deleted) SetStatus(val string) {
	s.Status = val
}

// ErrorStatusCode wraps Problem with StatusCode.
type ErrorStatusCode struct {
	StatusCode int
	Response   Problem
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorStatusCode) GetResponse() Problem {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorStatusCode) SetResponse(val Problem) {
	s.Response = val
}

// Ref: #/components/schemas/Problem
type Problem struct {
	Error string `json:"error"`
}

// GetError returns the value of Error.
func (s *Problem) GetError() string {
	return s.Error
}

// SetError sets the value of Error.
func (s *Problem) SetError(val string) {
	s.Error = val
}

// Merged schema.
// Ref: #/components/schemas/Publisher
type Publisher struct {
	Name string    `json:"name"`
	ID   uuid.UUID `json:"id"`
}

// GetName returns the value of Name.
func (s *Publisher) GetName() string {
	return s.Name
}

// GetID returns the value of ID.
func (s *Publisher) GetID() uuid.UUID {
	return s.ID
}

// SetName sets the value of Name.
func (s *Publisher) SetName(val string) {
	s.Name = val
}

// SetID sets the value of ID.
func (s *Publisher) SetID(val uuid.UUID) {
	s.ID = val
}

// Ref: #/components/schemas/PublisherInput
type PublisherInput struct {
	Name string `json:"name"`
}

// GetName returns the value of Name.
func (s *PublisherInput) GetName() string {
	return s.Name
}

// SetName sets the value of Name.
func (s *PublisherInput) SetName(val string) {
	s.Name = val
}